requirements.txt
requests==2.22.0
scikit-learn==0.21.3
numpy==1.17
google-cloud-storage==1.43.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu
torchvision==0.11.1+cpu
torchaudio==0.10.0+cpu
transformers==4.13.0
google-python-cloud-debugger